+++ packages/kiarina-lib-firebase-rtdb/src/kiarina/lib/firebase_rtdb/__init__.py
@@ -7,6 +7,7 @@ if TYPE_CHECKING:  # pragma: no cover
     from ._helpers.get_data import get_data
     from ._helpers.update_data import update_data
     from ._helpers.watch_data import watch_data
+    from ._models.rtdb_mirror import RTDBMirror
     from ._schemas.rtdb_query import RTDBQuery
     from ._settings import RTDBSettings, settings_manager
 
@@ -17,6 +18,8 @@ __all__ = [
     "get_data",
     "update_data",
     "watch_data",
+    # ._models
+    "RTDBMirror",
     # ._schemas
     "RTDBQuery",
     # ._settings
@@ -38,6 +41,8 @@ def __getattr__(name: str) -> object:
         "get_data": "._helpers.get_data",
         "update_data": "._helpers.update_data",
         "watch_data": "._helpers.watch_data",
+        # ._models
+        "RTDBMirror": "._models.rtdb_mirror",
         # ._schemas
         "RTDBQuery": "._schemas.rtdb_query",
         # ._settings